Multi-worker-docker-container: disable log buffering
Background: we have a `matrixdotorg/synapse-workers` docker image, which is
intended for running multiple workers within the same container. That image
includes a `prefix-log` script which, for each line printed to stdout or stderr
by one of the processes, prepends the name of the process.

This commit disables buffering in that script, so that lines are logged quickly
after they are printed. This makes it much easier to understand the output,
since they then come out in a natural order.