Include destinations we never sent to in catchup
When picking destination servers that we need to wake up for a retry, we need
to be mindful of destinations that we have *never* successfully sent to. This
can manifest either as a null `last_successful_stream_ordering`, or even no row
in `destinations` at all.

Hence, we need to left-join on `destinations` rather than inner-joining, and we
need to treat a null `last_successful_stream_ordering` the same as 0.