Patch the db conn pool sooner in tests
When running unit tests, we patch the database connection pool so that it runs
queries "synchronously". This is ok, except that if any queries are launched
before we do the patching, those queries get left in limbo and never complete.

To fix this, let's change the way we do the switcheroo, by patching out the
method which creates the connection pool in the first place.