Update the README with Element branding and a few fixes
This commit updates the Synapse README to make it clearer that this is
an Element-branded product and that people are not required, but have
the option to, pay Element for enterprise features and support.

Such support is critical to continued development of Synapse.

This commit also contains some aesthetic changes and a fix to the shield
icon that points to the community support room, which has been broken
for a long time. The breakage is due to it taking too long for
matrix.org to calculate the membership for that room on the endpoint
that shields.io requests. So it's been swapped out for a static shield.