Fix `_bulk_get_max_event_pos` being wrong/inefficient
It kept adding all of the `batch_results` to the `results`
over and over every time we checked a single room in the batch.

I think we still ended up with the right answer before because we accumulate
`recheck_rooms` and actually recheck them to overwrite the bad data
we wrote to the `results` before.