Fix saving of non-RGB thumbnails as PNG
The PNG format does not support non-RGB color spaces, which makes pillow
raise an exception when trying to save e.g. a CMYK image in PNG format.

Converting to RGB is recommended on the upstream issue, and as CMYK is a
print format transparency does not need to be considered.

Closes: #10741