Add `redirect_uri` option to `oidc_providers` entries
Allows overriding the `redirect_uri` parameter sent to both the
authorization and token endpoints of the IdP. Typically this parameter
is hardcoded to `<public_baseurl>/_synapse/client/oidc/callback`. But it
can be useful in certain contexts to allow a different callback URL. For
instance, if you would like to intercept the authorization code returned
from the IdP and do something with it, before eventually calling
Synapse's OIDC callback URL yourself.

This change enables enterprise use cases but does not change the default
behaviour.