Join policy server PoC
Adds the ability to enforce users to go through a web flow before being
able to join a room (unless invited).

Configured by specifying a "policy server" in the join rule event
content:

```
{
  "join_rule": "invite",
  "re.jki.join_policy_server": "localhost:8865"
}
```

The server will then return a 403 when a client tries to join, including
a URL that the client can redirect the user to, which eventually returns
a token (very much like an OAuth2 flow). This token then can be included
when calling `/join` again and the join will be successful.